Packaged code into duke,  duke.command and duke.item packages
