--- src/main/java/FormatOutput.java
@@ -1,7 +1,0 @@
-public class FormatOutput {
-    static String format(String s) {
-        return "    ____________________________________________________________\n     " +
-                s +
-                "\n    ____________________________________________________________\n";
-    }
-}
