+++ src/main/java/duke/Duke.java
@@ -1,6 +1,5 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.FormatStyle;
+package duke;
+
 import java.util.Scanner;
 
 public class Duke {
