+++ src/main/java/duke/Parser.java
@@ -1,9 +1,12 @@
+package duke;
+
+import duke.command.*;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 import java.time.format.DateTimeParseException;
 import java.time.format.FormatStyle;
-import java.util.ArrayList;
 import java.util.HashMap;
 
 public class Parser {
