+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.ItemList;
+import duke.CustomExceptions;
+
 public interface Command {
     String execute(String command, String[] info, ItemList itemList) throws CustomExceptions;
