+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,6 +1,11 @@
+package duke.command;
+
+import duke.ItemList;
+import duke.CustomExceptions;
+
 public class DeleteCommand implements Command {
     @Override
-    public String execute(String command, String[] info, ItemList itemList) throws CustomExceptions{
+    public String execute(String command, String[] info, ItemList itemList) throws CustomExceptions {
         if (info.length != 2) {
             throw new CustomExceptions.markException("Please enter delete command in the following format: delete <index>");
         } else {
