+++ src/main/java/duke/command/MarkCommand.java
@@ -1,7 +1,13 @@
+package duke.command;
+
+import duke.ItemList;
+import duke.CustomExceptions;
+import duke.item.Item;
+
 public class MarkCommand implements Command {
 
     @Override
-    public String execute(String command, String[] info, ItemList itemList) throws CustomExceptions{
+    public String execute(String command, String[] info, ItemList itemList) throws CustomExceptions {
         if (info.length != 2) {
             throw new CustomExceptions.markException("Please enter mark command in the following format: mark <index>");
         } else {
