+++ src/main/java/duke/item/Item.java
@@ -1,3 +1,5 @@
+package duke.item;
+
 import java.io.Serializable;
 
 public interface Item extends Serializable {
