+++ src/main/java/duke/item/ToDo.java
@@ -1,3 +1,5 @@
+package duke.item;
+
 import java.io.Serializable;
 
 public class ToDo implements Item, Serializable {
