+++ text-ui-test/EXPECTED.TXT
@@ -49,7 +49,7 @@
     ____________________________________________________________
 
     ____________________________________________________________
-     Please re-enter Deadline with a valid name.
+     Please re-enter duke.item.Deadline with a valid name.
     ____________________________________________________________
 
     ____________________________________________________________
@@ -57,7 +57,7 @@
     ____________________________________________________________
 
     ____________________________________________________________
-     Please re-enter Event with a valid name
+     Please re-enter duke.item.Event with a valid name
     ____________________________________________________________
 
     ____________________________________________________________
@@ -131,7 +131,7 @@
     ____________________________________________________________
 
     ____________________________________________________________
-     Please re-enter Deadline with a valid name
+     Please re-enter duke.item.Deadline with a valid name
     ____________________________________________________________
 
     ____________________________________________________________
@@ -139,11 +139,11 @@
     ____________________________________________________________
 
     ____________________________________________________________
-     Please re-enter Event /from BEFORE /to: event /to
+     Please re-enter duke.item.Event /from BEFORE /to: event /to
     ____________________________________________________________
 
     ____________________________________________________________
-     Please re-enter Event with a valid name
+     Please re-enter duke.item.Event with a valid name
     ____________________________________________________________
 
     ____________________________________________________________
@@ -157,7 +157,7 @@
     ____________________________________________________________
 
     ____________________________________________________________
-     Please re-enter Event with a valid name
+     Please re-enter duke.item.Event with a valid name
     ____________________________________________________________
 
     ____________________________________________________________
